refactor and javadoc
